requirements.txt
# Core Flask
Flask==3.1.1
Flask-CORS==5.0.1
Flask-Compress==1.18
Flask-Limiter==3.12
Werkzeug==3.1.3

# Fast JSON
orjson==3.11.2

# Production server
gunicorn==23.0.0

# Bitcoin/Crypto
secp256k1==0.14.0
base58==2.1.1
ecdsa==0.19.1

# Dependencies
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.0
cffi==1.17.1
pycparser==2.22
six==1.17.0

# Rate limiting dependencies
limits==5.5.0
Deprecated==1.2.18
wrapt==1.17.3
typing_extensions==4.15.0
packaging==25.0
ordered-set==4.1.0
redis==6.4.0

# Compression dependencies  
Brotli==1.1.0

# Optional (if needed for other parts)
requests==2.32.3
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0

# Testing utilities
hypothesis==6.140.2
pytest-cov==7.0.0
pytest-env==1.1.5
